refactor(meet): simplify E2EE media plumbing
